requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
mysqlclient==2.2.0
Flask-WTF==1.2.1
email-validator==2.0.0
